clean up table ui; actions & badge style
