--- packages/web/src/pages/_protected/groups/$slug/@test-dialog.tsx
@@ -1,63 +1,0 @@
-import { Button } from "@/components/ui/button";
-
-import {
-  Select,
-  SelectContent,
-  SelectGroup,
-  SelectItem,
-  SelectLabel,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import {
-  Sheet,
-  SheetContent,
-  SheetTitle,
-  SheetDescription,
-  SheetHeader,
-  SheetTrigger,
-  SheetFooter,
-} from "@/components/ui/sheet";
-
-export function Test() {
-  return (
-    <Sheet>
-      <SheetTrigger asChild>
-        <Button variant="outline">Edit Profile</Button>
-      </SheetTrigger>
-      <SheetContent className="sm:max-w-[425px]">
-        <form action="">
-          <SheetHeader>
-            <SheetTitle>Edit profile</SheetTitle>
-            <SheetDescription>
-              Make changes to your profile here. Click save when you're done.
-            </SheetDescription>
-          </SheetHeader>
-          <Select>
-            <SelectTrigger className="w-[280px]">
-              <SelectValue placeholder="Select a timezone" />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectGroup>
-                <SelectLabel>North America</SelectLabel>
-                <SelectItem value="est">Eastern Standard Time (EST)</SelectItem>
-                <SelectItem value="cst">Central Standard Time (CST)</SelectItem>
-                <SelectItem value="mst">
-                  Mountain Standard Time (MST)
-                </SelectItem>
-                <SelectItem value="pst">Pacific Standard Time (PST)</SelectItem>
-                <SelectItem value="akst">
-                  Alaska Standard Time (AKST)
-                </SelectItem>
-                <SelectItem value="hst">Hawaii Standard Time (HST)</SelectItem>
-              </SelectGroup>
-            </SelectContent>
-          </Select>
-          <SheetFooter>
-            <Button type="submit">Save changes</Button>
-          </SheetFooter>
-        </form>
-      </SheetContent>
-    </Sheet>
-  );
-}
